[hotfix] Let delayed AdaptiveScheduler.runIfState method return ScheduledFuture
By letting the AdaptiveScheduler.runIfState method return a ScheduledFuture it is now
possible to cancel scheduled actions.
